- wxScintilla: Fix narrowing initialization
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@11032 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/wxscintilla/src/ScintillaWX.cpp
@@ -290,7 +290,7 @@ static int wxCountLines(const char* text, int scintillaMode)
 
 /* C::B begin */
 // Constant ids for the timers used by every editor.
-const int timerIDs[ScintillaWX::timersCount] = { wxNewId(), wxNewId(), wxNewId(), wxNewId() };
+const long int timerIDs[ScintillaWX::timersCount] = { wxNewId(), wxNewId(), wxNewId(), wxNewId() };
 /* C::B end */
 
 //----------------------------------------------------------------------
